Cambiada vista basada en función por vista basada en clase para planilla de movimientos. Retiradas vistas de función que ya no se usan de finper.views (se guardaron en finper.obsolete.views_obsolete)
