+++ djangofinper/templates/base.html
@@ -8,11 +8,12 @@
         {{ page_title|default:"Finanzas Personales" }}
         {% endblock title %}
     </title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
     <link href="{% static 'main.css' %}" rel="stylesheet" type="text/css">
 </head>
 
 <body>
-    <div id="content">
+    <div class="container-fluid" id="content">
         {% block content %}
         <p>Placeholder text</p>
         {% endblock content %}
